src/main/java/kr/or/ddit/db/CustomSqlSessionFactoryBuilder.java
package kr.or.ddit.db;

import java.io.IOException;
import java.io.Reader;

import org.apache.ibatis.io.Resources;
import org.apache.ibatis.session.SqlSessionFactory;
import org.apache.ibatis.session.SqlSessionFactoryBuilder;

/**
 * sqlSession : 지역변수 형태의 메소드 단위로 관리.
 * sqlSessionFactory : 싱글턴 객체로 관리.
 *
 */
public class CustomSqlSessionFactoryBuilder {
	private static SqlSessionFactory sqlSessionFactory;
	static {
		String configFile = "kr/or/ddit/mybatis/Configuration.xml";
		try(
			Reader reader =  Resources.getResourceAsReader(configFile);				
		) {
			sqlSessionFactory = new SqlSessionFactoryBuilder().build(reader);
		} catch (IOException e) {
			throw new RuntimeException(e);
		}
	}
	
	public static SqlSessionFactory getSqlSessionFactory() {
		return sqlSessionFactory;
	}
}
